Refactor into one lock
