+++ src/CacheTower.Extensions.Redis/RedisLockExtension.cs
@@ -93,7 +93,38 @@ namespace CacheTower.Extensions.Redis
 					var tcs = new TaskCompletionSource<bool>();
 					var cts = new CancellationTokenSource(Options.LockTimeout);
 					cts.Token.Register(tcs => ((TaskCompletionSource<bool>)tcs).TrySetCanceled(), tcs, useSynchronizationContext: false);
+
+					if (Options.UseBusyLockCheck)
+					{
+						_ = TestLock(tcs);
+					}
+
 					return tcs;
+
+					async Task TestLock(TaskCompletionSource<bool> taskCompletionSource)
+					{
+						var spinAttempt = 0;
+
+						while (spinAttempt <= Options.SpinAttempts && 
+						       !taskCompletionSource.Task.IsCanceled && 
+						       !taskCompletionSource.Task.IsCompleted)
+						{
+							spinAttempt++;
+
+							var lockQuery = await Database.LockQueryAsync(lockKey);
+
+							if (lockQuery.HasValue)
+							{
+								await Task.Delay(Options.SpinTime);
+								continue;
+							}
+
+							taskCompletionSource.TrySetResult(true);
+							return;
+						}
+
+						taskCompletionSource.TrySetCanceled();
+					}
 				});
 
 				//Last minute check to confirm whether waiting is required (in case the notification is missed)
